+++ backend/src/main/java/com/pingcap/ecommerce/dao/tidb/OrderMapper.java
@@ -8,6 +8,7 @@ import org.apache.ibatis.annotations.Mapper;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Mapper
@@ -22,7 +23,7 @@ public interface OrderMapper {
 
     List<Long> getOrderIds(Pageable pageable);
 
-    long getOrdersCount(String userId);
+    BigInteger getOrdersCount(String userId);
 
     List<OrderVO> getOrdersByUserId(String userId);
